moved bunch of stuff from App.js to Foodlist compontent :rocket::moon:
